jobs/gfs_snow_tail/requirements.txt
google-cloud-storage==2.17.0
requests==2.32.3
